requirements.txt
langchain
langchain-community
langchain-classic
langchain-groq
langchain-ollama
langchain-huggingface
langchain-text-splitters
python-dotenv
sentence-transformers
pypdf
beautifulsoup4
lxml
hf_xet
sse-starlette
streamlit
faiss-cpu
fpdf
